revised migration user phone to bigIntiger for deployment

+++ data/migrations/20201014132916_createtables.js
@@ -4,7 +4,7 @@ exports.up = function (knex) {
         .createTable('users', tbl => {
             tbl.increments();
             tbl.string('username', 128).notNullable().unique();
-            tbl.integer('phone_number');
+            tbl.bigInteger('phone_number');
             tbl.varchar('password', 128).notNullable();
         })
         .createTable('plants', tbl => {
